Moved `ghost` helper to `helpers`
no refs

Moved the `ghost` helper to `helpers` and updated all references to it
