+++ src/helpers/ghost.unit.test.ts
@@ -9,7 +9,7 @@ import {
     ACTOR_DEFAULT_ICON,
     ACTOR_DEFAULT_NAME,
     ACTOR_DEFAULT_SUMMARY
-} from './constants';
+} from '../constants';
 
 describe('getSiteSettings', function () {
     const host = 'example.com';
